load_data refactoring proposal: file structure and method refactoring
