+++ api/scpca_portal/models/project.py
@@ -64,66 +64,9 @@ class Project(CommonDataAttributes, TimestampedModel):
     def __str__(self):
         return f"Project {self.scpca_id}"
 
-    @staticmethod
-    def get_input_project_metadata_file_path():
-        return common.INPUT_DATA_PATH / "project_metadata.csv"
-
     @property
     def computed_files(self):
         return self.project_computed_files.order_by("created_at")
-
-    @property
-    def input_data_path(self):
-        return common.INPUT_DATA_PATH / self.scpca_id
-
-    @property
-    def input_bulk_metadata_file_path(self):
-        return self.input_data_path / f"{self.scpca_id}_bulk_metadata.tsv"
-
-    @property
-    def input_bulk_quant_file_path(self):
-        return self.input_data_path / f"{self.scpca_id}_bulk_quant.tsv"
-
-    @property
-    def input_samples_metadata_file_path(self):
-        return self.input_data_path / "samples_metadata.csv"
-
-    @property
-    def multiplexed_computed_file(self):
-        try:
-            return self.project_computed_files.get(
-                type=ComputedFile.OutputFileTypes.PROJECT_MULTIPLEXED_ZIP
-            )
-        except ComputedFile.DoesNotExist:
-            pass
-
-    @property
-    def output_multiplexed_computed_file_name(self):
-        return f"{self.scpca_id}_multiplexed.zip"
-
-    @property
-    def output_multiplexed_metadata_file_path(self):
-        return common.OUTPUT_DATA_PATH / f"{self.scpca_id}_multiplexed_metadata.tsv"
-
-    @property
-    def output_single_cell_computed_file_name(self):
-        return f"{self.scpca_id}.zip"
-
-    @property
-    def output_single_cell_anndata_computed_file_name(self):
-        return f"{self.scpca_id}_anndata.zip"
-
-    @property
-    def output_single_cell_metadata_file_path(self):
-        return common.OUTPUT_DATA_PATH / f"{self.scpca_id}_libraries_metadata.tsv"
-
-    @property
-    def output_spatial_computed_file_name(self):
-        return f"{self.scpca_id}_spatial.zip"
-
-    @property
-    def output_spatial_metadata_file_path(self):
-        return common.OUTPUT_DATA_PATH / f"{self.scpca_id}_spatial_metadata.tsv"
 
     @property
     def single_cell_computed_file(self):
@@ -135,6 +78,7 @@ class Project(CommonDataAttributes, TimestampedModel):
         except ComputedFile.DoesNotExist:
             pass
 
+    # FOR TESTING
     @property
     def single_cell_anndata_computed_file(self):
         try:
@@ -155,150 +99,14 @@ class Project(CommonDataAttributes, TimestampedModel):
             pass
 
     @property
-    def url(self):
-        return f"https://scpca.alexslemonade.org/projects/{self.scpca_id}"
-
-    def add_project_metadata(self, sample_metadata):
-        """Adds project level metadata to the `sample_metadata`."""
-        sample_metadata["pi_name"] = self.pi_name
-        sample_metadata["project_title"] = self.title
-        sample_metadata["scpca_project_id"] = self.scpca_id
-
-    def combine_multiplexed_metadata(
-        self,
-        samples_metadata: List[Dict],
-        multiplexed_libraries_metadata: List[Dict],
-        sample_id: str,
-    ):
-        """Combines the two metadata dicts together to have all multiplexed data
-        at the library level. Writes the combination out to the project and
-        sample metadata files. The sample file also includes multiplexed samples.
-        """
-
-        combined_metadata = []
-        multiplexed_sample_mapping = {}
-        if not multiplexed_libraries_metadata:
-            return combined_metadata, multiplexed_sample_mapping
-
-        modality = Sample.Modalities.MULTIPLEXED
-        library_metadata_keys = self.get_library_metadata_keys(
-            set(multiplexed_libraries_metadata[0].keys()), modalities={modality}
-        )
-        sample_metadata_keys = self.get_sample_metadata_keys(
-            set(samples_metadata[0].keys()), modalities={modality}
-        )
-        field_names = self.get_metadata_field_names(
-            library_metadata_keys.union(sample_metadata_keys), modality=modality
-        )
-
-        multiplexed_library_mapping = {}  # Sample ID to library IDs mapping.
-        multiplexed_sample_ids = set()  # Unified multiplexed sample ID set.
-        for library_metadata in multiplexed_libraries_metadata:
-            multiplexed_library_sample_ids = library_metadata["demux_samples"]
-            for multiplexed_sample_id in multiplexed_library_sample_ids:
-                # Populate multiplexed library mapping.
-                if multiplexed_sample_id not in multiplexed_library_mapping:
-                    multiplexed_library_mapping[multiplexed_sample_id] = set()
-                multiplexed_library_mapping[multiplexed_sample_id].add(
-                    library_metadata["scpca_library_id"]
-                )
-
-                # Add sample IDs to a unified set.
-                multiplexed_sample_ids.update(multiplexed_library_sample_ids)
-
-                # Remove sample ID from a mapping as sample cannot be
-                # multiplexed with itself.
-                multiplexed_library_sample_ids_copy = set(multiplexed_library_sample_ids)
-                multiplexed_library_sample_ids_copy.discard(multiplexed_sample_id)
-
-                # Populate multiplexed sample mapping.
-                if multiplexed_sample_id not in multiplexed_sample_mapping:
-                    multiplexed_sample_mapping[multiplexed_sample_id] = set()
-                multiplexed_sample_mapping[multiplexed_sample_id].update(
-                    multiplexed_library_sample_ids_copy
-                )
-
-        # Generate multiplexed sample metadata dict.
-        sample_metadata_mapping = {}
-        for sample_metadata in samples_metadata:
-            multiplexed_sample_id = sample_metadata["scpca_sample_id"]
-            if multiplexed_sample_id not in multiplexed_sample_ids:  # Skip non-multiplexed samples.
-                continue
-
-            if sample_id and multiplexed_sample_id != sample_id:
-                continue
-
-            sample_metadata_copy = sample_metadata.copy()
-            for key in sample_metadata.keys():  # Exclude fields.
-                if key not in sample_metadata_keys:
-                    sample_metadata_copy.pop(key)
-
-            self.add_project_metadata(sample_metadata_copy)
-            sample_metadata_mapping[multiplexed_sample_id] = sample_metadata_copy
-
-        # Combine and write the metadata.
-        combined_metadata_added_pair_ids = set()
-        for sample_id in sorted(sample_metadata_mapping.keys()):
-            sample_metadata_path = Sample.get_output_metadata_file_path(sample_id, modality)
-            with open(sample_metadata_path, "w", newline="") as sample_file:
-                sample_csv_writer = csv.DictWriter(
-                    sample_file, fieldnames=field_names, delimiter=common.TAB
-                )
-                sample_csv_writer.writeheader()
-
-                multiplexed_sample_ids = sorted(multiplexed_sample_mapping[sample_id])
-                multiplexed_sample_ids.insert(0, sample_id)  # Current sample libraries go first.
-                for multiplexed_sample_id in multiplexed_sample_ids:
-                    libraries_metadata = sorted(
-                        (
-                            library
-                            for library in multiplexed_libraries_metadata
-                            if library["scpca_library_id"] in multiplexed_library_mapping[sample_id]
-                        ),
-                        key=lambda library: library["scpca_library_id"],
-                    )
-                    for library_metadata in libraries_metadata:
-                        library_metadata_copy = library_metadata.copy()
-                        for key in library_metadata.keys():  # Exclude fields.
-                            if key not in library_metadata_keys:
-                                library_metadata_copy.pop(key)
-
-                        library_metadata_copy.update(
-                            sample_metadata_mapping.get(multiplexed_sample_id, {})
-                        )
-                        sample_csv_writer.writerow(library_metadata_copy)
-
-                        pair_id = (library_metadata_copy["scpca_library_id"], multiplexed_sample_id)
-                        if pair_id not in combined_metadata_added_pair_ids:
-                            combined_metadata_added_pair_ids.add(pair_id)
-                            combined_metadata.append(library_metadata_copy)
-
-        with open(self.output_multiplexed_metadata_file_path, "w", newline="") as project_file:
-            project_csv_writer = csv.DictWriter(
-                project_file, fieldnames=field_names, delimiter=common.TAB
+    def multiplexed_computed_file(self):
+        try:
+            return self.project_computed_files.get(
+                type=ComputedFile.OutputFileTypes.PROJECT_MULTIPLEXED_ZIP
             )
-            project_csv_writer.writeheader()
-            # Project file data has to be sorted by the library_id.
-            project_csv_writer.writerows(
-                sorted([cm for cm in combined_metadata], key=lambda cm: cm["scpca_library_id"])
-            )
-
-        return combined_metadata, multiplexed_sample_mapping
-
-    @staticmethod
-    def process_computed_file(computed_file, clean_up_output_data, update_s3):
-        """Processes saving, upload and cleanup of a single computed file."""
-        logger.info(f"Processing {computed_file}")
-
-        computed_file.save()
-        if update_s3:
-            logger.info(f"Uploading {computed_file}")
-            computed_file.create_s3_file()
-        if clean_up_output_data:
-            computed_file.zip_file_path.unlink(missing_ok=True)
-
-        # Close DB connection for each thread.
-        connection.close()
+        except ComputedFile.DoesNotExist:
+            pass
+    # END
 
     def add_contacts(self, contact_email, contact_name):
         """Creates and adds project contacts."""
@@ -365,445 +173,6 @@ class Project(CommonDataAttributes, TimestampedModel):
 
             self.publications.add(publication)
 
-    def combine_single_cell_metadata(
-        self,
-        samples_metadata: List[Dict],
-        single_cell_libraries_metadata: List[Dict],
-        sample_id: str,
-    ):
-        """Combines the two metadata dicts together to have all single cell data
-        at the library level. Writes the combination out to the project and
-        sample metadata files.
-        """
-        combined_metadata = []
-
-        if not single_cell_libraries_metadata:
-            return combined_metadata
-
-        modality = Sample.Modalities.SINGLE_CELL
-        modalities = {modality}
-        if self.has_cite_seq_data:
-            modalities.add(Sample.Modalities.CITE_SEQ)
-
-        library_metadata_keys = self.get_library_metadata_keys(
-            set(single_cell_libraries_metadata[0].keys()), modalities=modalities
-        )
-        sample_metadata_keys = self.get_sample_metadata_keys(
-            set(samples_metadata[0].keys()), modalities=modalities
-        )
-        field_names = self.get_metadata_field_names(
-            library_metadata_keys.union(sample_metadata_keys), modality=modality
-        )
-
-        with open(self.output_single_cell_metadata_file_path, "w", newline="") as project_file:
-            project_csv_writer = csv.DictWriter(
-                project_file, fieldnames=field_names, delimiter=common.TAB
-            )
-            project_csv_writer.writeheader()
-
-            for sample_metadata in samples_metadata:
-                scpca_sample_id = sample_metadata["scpca_sample_id"]
-                if sample_id and scpca_sample_id != sample_id:
-                    continue
-
-                sample_metadata_copy = sample_metadata.copy()
-                for key in sample_metadata.keys():  # Exclude fields.
-                    if key not in sample_metadata_keys:
-                        sample_metadata_copy.pop(key)
-
-                self.add_project_metadata(sample_metadata_copy)
-
-                sample_metadata_path = Sample.get_output_metadata_file_path(
-                    scpca_sample_id, modality
-                )
-                with open(sample_metadata_path, "w", newline="") as sample_file:
-                    sample_csv_writer = csv.DictWriter(
-                        sample_file, fieldnames=field_names, delimiter=common.TAB
-                    )
-                    sample_csv_writer.writeheader()
-
-                    libraries_metadata = (
-                        library
-                        for library in single_cell_libraries_metadata
-                        if library["scpca_sample_id"] == scpca_sample_id
-                    )
-                    for library_metadata in libraries_metadata:
-                        library_metadata_copy = library_metadata.copy()
-                        for key in library_metadata.keys():  # Exclude fields.
-                            if key not in library_metadata_keys:
-                                library_metadata_copy.pop(key)
-
-                        library_metadata_copy.update(sample_metadata_copy)
-                        combined_metadata.append(library_metadata_copy)
-
-                        sample_csv_writer.writerow(library_metadata_copy)
-                        project_csv_writer.writerow(library_metadata_copy)
-
-        return combined_metadata
-
-    def combine_spatial_metadata(
-        self,
-        samples_metadata: List[Dict],
-        spatial_libraries_metadata: List[Dict],
-        sample_id: str,
-    ):
-        """Combines the two metadata dicts together to have all spatial data at
-        the library level. Writes the combination out to the project and
-        sample metadata files.
-        """
-        combined_metadata = []
-
-        if not spatial_libraries_metadata:
-            return combined_metadata
-
-        modality = Sample.Modalities.SPATIAL
-        library_metadata_keys = self.get_library_metadata_keys(
-            set(spatial_libraries_metadata[0].keys()), modalities={modality}
-        )
-        sample_metadata_keys = self.get_sample_metadata_keys(
-            set(samples_metadata[0].keys()), modalities={modality}
-        )
-        field_names = self.get_metadata_field_names(
-            library_metadata_keys.union(sample_metadata_keys), modality=modality
-        )
-
-        with open(self.output_spatial_metadata_file_path, "w", newline="") as project_file:
-            project_csv_writer = csv.DictWriter(
-                project_file, fieldnames=field_names, delimiter=common.TAB
-            )
-            project_csv_writer.writeheader()
-
-            for sample_metadata in samples_metadata:
-                scpca_sample_id = sample_metadata["scpca_sample_id"]
-                if sample_id and scpca_sample_id != sample_id:
-                    continue
-
-                sample_metadata_copy = sample_metadata.copy()
-                for key in sample_metadata.keys():  # Exclude fields.
-                    if key not in sample_metadata_keys:
-                        sample_metadata_copy.pop(key)
-
-                self.add_project_metadata(sample_metadata_copy)
-
-                sample_metadata_path = Sample.get_output_metadata_file_path(
-                    scpca_sample_id, modality
-                )
-                with open(sample_metadata_path, "w", newline="") as sample_file:
-                    sample_csv_writer = csv.DictWriter(
-                        sample_file, fieldnames=field_names, delimiter=common.TAB
-                    )
-                    sample_csv_writer.writeheader()
-
-                    libraries_metadata = (
-                        library
-                        for library in spatial_libraries_metadata
-                        if library["scpca_sample_id"] == scpca_sample_id
-                    )
-                    for library_metadata in libraries_metadata:
-                        library_metadata_copy = library_metadata.copy()
-                        for key in library_metadata.keys():  # Exclude fields.
-                            if key not in library_metadata_keys:
-                                library_metadata_copy.pop(key)
-
-                        library_metadata_copy.update(sample_metadata_copy)
-                        combined_metadata.append(library_metadata_copy)
-
-                        sample_csv_writer.writerow(library_metadata_copy)
-                        project_csv_writer.writerow(library_metadata_copy)
-
-        return combined_metadata
-
-    def create_anndata_readme_file(self):
-        """Creates a annotation metadata README file."""
-        with open(ComputedFile.README_ANNDATA_FILE_PATH, "w") as readme_file:
-            readme_file.write(
-                render_to_string(
-                    ComputedFile.README_TEMPLATE_ANNDATA_FILE_PATH,
-                    context={
-                        "additional_terms": self.get_additional_terms(),
-                        "date": utils.get_today_string(),
-                        "project_accession": self.scpca_id,
-                        "project_url": self.url,
-                    },
-                ).strip()
-            )
-
-    def create_single_cell_readme_file(self):
-        """Creates a single cell metadata README file."""
-        with open(ComputedFile.README_SINGLE_CELL_FILE_PATH, "w") as readme_file:
-            readme_file.write(
-                render_to_string(
-                    ComputedFile.README_TEMPLATE_SINGLE_CELL_FILE_PATH,
-                    context={
-                        "additional_terms": self.get_additional_terms(),
-                        "date": utils.get_today_string(),
-                        "project_accession": self.scpca_id,
-                        "project_url": self.url,
-                    },
-                ).strip()
-            )
-
-    def create_multiplexed_readme_file(self):
-        """Creates a multiplexed metadata README file."""
-        with open(ComputedFile.README_MULTIPLEXED_FILE_PATH, "w") as readme_file:
-            readme_file.write(
-                render_to_string(
-                    ComputedFile.README_TEMPLATE_MULTIPLEXED_FILE_PATH,
-                    context={
-                        "additional_terms": self.get_additional_terms(),
-                        "date": utils.get_today_string(),
-                        "project_accession": self.scpca_id,
-                        "project_url": self.url,
-                    },
-                ).strip()
-            )
-
-    def create_spatial_readme_file(self):
-        """Creates a spatial metadata README file."""
-        with open(ComputedFile.README_SPATIAL_FILE_PATH, "w") as readme_file:
-            readme_file.write(
-                render_to_string(
-                    ComputedFile.README_TEMPLATE_SPATIAL_FILE_PATH,
-                    context={
-                        "additional_terms": self.get_additional_terms(),
-                        "date": utils.get_today_string(),
-                        "project_accession": self.scpca_id,
-                        "project_url": self.url,
-                    },
-                ).strip()
-            )
-
-    def get_bulk_rna_seq_sample_ids(self):
-        """Returns bulk RNA sequencing sample IDs."""
-        bulk_rna_seq_sample_ids = set()
-        if self.has_bulk_rna_seq:
-            with open(self.input_bulk_metadata_file_path, "r") as bulk_metadata_file:
-                bulk_rna_seq_sample_ids.update(
-                    (
-                        line["sample_id"]
-                        for line in csv.DictReader(bulk_metadata_file, delimiter=common.TAB)
-                    )
-                )
-        return bulk_rna_seq_sample_ids
-
-    def create_computed_files(
-        self,
-        single_cell_file_mapping,
-        single_cell_workflow_versions,
-        spatial_file_mapping,
-        spatial_workflow_versions,
-        multiplexed_file_mapping,
-        multiplexed_workflow_versions,
-        max_workers=6,  # 6 = 2 file formats * 3 mappings.
-        clean_up_output_data=True,
-        update_s3=False,
-    ):
-        """Prepares ready for saving project computed files based on generated file mappings."""
-
-        def create_computed_file(future):
-            computed_file = future.result()
-            self.process_computed_file(computed_file, clean_up_output_data, update_s3)
-
-        with ThreadPoolExecutor(max_workers=max_workers) as tasks:
-            for file_format in (
-                ComputedFile.OutputFileFormats.ANN_DATA,
-                ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT,
-            ):
-                if multiplexed_file_mapping.get(file_format):
-                    tasks.submit(
-                        ComputedFile.get_project_multiplexed_file,
-                        self,
-                        multiplexed_file_mapping,
-                        multiplexed_workflow_versions,
-                        file_format,
-                    ).add_done_callback(create_computed_file)
-
-                if single_cell_file_mapping.get(file_format):
-                    tasks.submit(
-                        ComputedFile.get_project_single_cell_file,
-                        self,
-                        single_cell_file_mapping,
-                        single_cell_workflow_versions,
-                        file_format,
-                    ).add_done_callback(create_computed_file)
-
-                if spatial_file_mapping.get(file_format):
-                    tasks.submit(
-                        ComputedFile.get_project_spatial_file,
-                        self,
-                        spatial_file_mapping,
-                        spatial_workflow_versions,
-                        file_format,
-                    ).add_done_callback(create_computed_file)
-
-    def get_additional_terms(self):
-        if not self.additional_restrictions:
-            return ""
-
-        with open(
-            common.TEMPLATE_PATH / "readme/additional_terms/research_academic_only.md"
-        ) as additional_terms_file:
-            return additional_terms_file.read()
-
-    def get_library_metadata_keys(self, all_keys, modalities=()):
-        """Returns a set of library metadata keys based on the modalities context."""
-        excluded_keys = {
-            "scpca_sample_id",
-        }
-
-        if Sample.Modalities.CITE_SEQ not in modalities:
-            excluded_keys.add("has_citeseq")
-
-        if Sample.Modalities.SPATIAL in modalities:
-            excluded_keys.update(
-                (
-                    "filtered_cells",
-                    "filtered_spots",
-                    "tissue_spots",
-                    "unfiltered_cells",
-                    "unfiltered_spots",
-                )
-            )
-
-        return all_keys.difference(excluded_keys)
-
-    def get_metadata_field_names(self, columns, modality):
-        """Returns a list of metadata field names based on the modality context."""
-        ordering = {
-            Sample.Modalities.MULTIPLEXED: (
-                "scpca_sample_id",
-                "scpca_library_id",
-                "scpca_project_id",
-                "technology",
-                "seq_unit",
-                "total_reads",
-                "mapped_reads",
-                "genome_assembly",
-                "mapping_index",
-                "date_processed",
-                "spaceranger_version",
-                "workflow",
-                "workflow_version",
-                "workflow_commit",
-                "diagnosis",
-                "subdiagnosis",
-                "pi_name",
-                "project_title",
-                "disease_timing",
-                "age_at_diagnosis",
-                "sex",
-                "tissue_location",
-                "treatment",
-                "participant_id",
-                "submitter",
-                "submitter_id",
-            ),
-            Sample.Modalities.SINGLE_CELL: (
-                "scpca_sample_id",
-                "scpca_library_id",
-                "diagnosis",
-                "subdiagnosis",
-                "seq_unit",
-                "technology",
-                "sample_cell_count_estimate",
-                "scpca_project_id",
-                "pi_name",
-                "project_title",
-                "disease_timing",
-                "age_at_diagnosis",
-                "sex",
-                "tissue_location",
-            ),
-            Sample.Modalities.SPATIAL: (
-                "scpca_project_id",
-                "scpca_sample_id",
-                "scpca_library_id",
-                "technology",
-                "seq_unit",
-                "total_reads",
-                "mapped_reads",
-                "genome_assembly",
-                "mapping_index",
-                "date_processed",
-                "spaceranger_version",
-                "workflow",
-                "workflow_version",
-                "workflow_commit",
-                "diagnosis",
-                "subdiagnosis",
-                "pi_name",
-                "project_title",
-                "disease_timing",
-                "age_at_diagnosis",
-                "sex",
-                "tissue_location",
-                "treatment",
-                "participant_id",
-                "submitter",
-                "submitter_id",
-            ),
-        }
-
-        return sorted(
-            sorted((c for c in columns), key=str.lower),  # Sort by a column name first.
-            key=lambda k: (
-                ordering[modality].index(k)  # Then enforce expected ordering.
-                if k in ordering[modality]
-                else float("inf")
-            ),
-        )
-
-    def get_sample_metadata_keys(self, all_keys, modalities=()):
-        """Returns a set of metadata keys based on the modalities context."""
-        excluded_keys = {
-            "has_bulk_rna_seq",
-            "has_cite_seq_data",
-            "has_single_cell_data",
-            "has_spatial_data",
-            "seq_units",
-            "technologies",
-        }
-        project_keys = {
-            "pi_name",
-            "project_title",
-        }
-
-        if Sample.Modalities.MULTIPLEXED in modalities:
-            excluded_keys.update(
-                (
-                    "alevin_fry_version",
-                    "date_processed",
-                    "filtered_cell_count",
-                    "mapped_reads",
-                    "scpca_library_id",
-                    "sample_cell_count_estimate",
-                    "total_reads",
-                    "workflow",
-                    "workflow_commit",
-                    "workflow_version",
-                )
-            )
-
-        if Sample.Modalities.SPATIAL in modalities:
-            excluded_keys.update(
-                (
-                    "alevin_fry_version",
-                    "filtered_cell_count",
-                    "filtering_method",
-                    "salmon_version",
-                    "sample_cell_count_estimate",
-                    "transcript_type",
-                    "unfiltered_cells",
-                    "workflow_version",
-                )
-            )
-
-        return all_keys.union(project_keys).difference(excluded_keys)
-
-    def get_sample_input_data_dir(self, sample_scpca_id):
-        """Returns an input data directory based on a sample ID."""
-        return self.input_data_path / sample_scpca_id
-
     def get_samples(
         self,
         samples_metadata,
@@ -847,298 +216,7 @@ class Project(CommonDataAttributes, TimestampedModel):
 
         return samples
 
-    def load_data(self, sample_id=None, **kwargs) -> None:
-        """
-        Goes through a project directory contents, parses multiple level metadata
-        files, writes combined metadata into resulting files.
-
-        Returns a list of project's computed files.
-        """
-
-        # Start with a list of samples and their metadata.
-        with open(self.input_samples_metadata_file_path) as samples_csv_file:
-            samples_metadata = [line for line in csv.DictReader(samples_csv_file)]
-
-        self.create_anndata_readme_file()
-        self.create_multiplexed_readme_file()
-        self.create_single_cell_readme_file()
-        self.create_spatial_readme_file()
-
-        bulk_rna_seq_sample_ids = self.get_bulk_rna_seq_sample_ids()
-        non_downloadable_sample_ids = set()
-        single_cell_libraries_metadata = []
-        spatial_libraries_metadata = []
-        for sample_metadata in samples_metadata:
-            scpca_sample_id = sample_metadata["scpca_sample_id"]
-            if sample_id and scpca_sample_id != sample_id:
-                continue
-
-            # Some samples will exist but their contents cannot be shared yet.
-            # When this happens their corresponding sample folder will not exist.
-            sample_dir = self.get_sample_input_data_dir(scpca_sample_id)
-            if not sample_dir.exists():
-                non_downloadable_sample_ids.add(scpca_sample_id)
-
-            has_cite_seq_data = False
-            has_single_cell_data = False
-            has_spatial_data = False
-            sample_cell_count_estimate = 0
-            sample_seq_units = set()
-            sample_technologies = set()
-            # Handle single cell metadata.
-            for filename_path in sorted(Path(sample_dir).glob("*_metadata.json")):
-                with open(filename_path) as sample_json_file:
-                    single_cell_json = json.load(sample_json_file)
-
-                has_single_cell_data = True
-                # Some rare samples can have one library with CITE-seq data and another library
-                # next to it without CITE-seq data (e.g. SCPCP000008/SCPCS000368).
-                has_cite_seq_data = single_cell_json.get("has_citeseq", False) or has_cite_seq_data
-
-                single_cell_json["filtered_cell_count"] = single_cell_json.pop("filtered_cells")
-                single_cell_json["scpca_library_id"] = single_cell_json.pop("library_id")
-                single_cell_json["scpca_sample_id"] = single_cell_json.pop("sample_id")
-                single_cell_libraries_metadata.append(single_cell_json)
-
-                sample_cell_count_estimate += single_cell_json["filtered_cell_count"]
-                sample_seq_units.add(single_cell_json["seq_unit"].strip())
-                sample_technologies.add(single_cell_json["technology"].strip())
-
-            # Handle spatial metadata.
-            for filename_path in sorted(Path(sample_dir).rglob("*_spatial/*_metadata.json")):
-                with open(filename_path) as spatial_json_file:
-                    spatial_json = json.load(spatial_json_file)
-                has_spatial_data = True
-
-                spatial_json["scpca_library_id"] = spatial_json.pop("library_id")
-                spatial_json["scpca_sample_id"] = spatial_json.pop("sample_id")
-                spatial_libraries_metadata.append(spatial_json)
-
-                sample_seq_units.add(spatial_json["seq_unit"].strip())
-                sample_technologies.add(spatial_json["technology"].strip())
-
-            sample_metadata["age_at_diagnosis"] = sample_metadata.pop("age")
-            sample_metadata["has_bulk_rna_seq"] = scpca_sample_id in bulk_rna_seq_sample_ids
-            sample_metadata["has_cite_seq_data"] = has_cite_seq_data
-            sample_metadata["has_single_cell_data"] = has_single_cell_data
-            sample_metadata["has_spatial_data"] = has_spatial_data
-            sample_metadata["includes_anndata"] = len(list(Path(sample_dir).glob("*.hdf5"))) > 0
-            sample_metadata["sample_cell_count_estimate"] = sample_cell_count_estimate
-            sample_metadata["seq_units"] = ", ".join(sorted(sample_seq_units, key=str.lower))
-            sample_metadata["technologies"] = ", ".join(sorted(sample_technologies, key=str.lower))
-
-        multiplexed_libraries_metadata = []
-        multiplexed_library_path_mapping = {}
-        multiplexed_sample_demux_cell_counter = Counter()
-        multiplexed_sample_seq_units_mapping = {}
-        multiplexed_sample_technologies_mapping = {}
-        for multiplexed_sample_dir in sorted(Path(self.input_data_path).rglob("*,*")):
-            for filename_path in sorted(Path(multiplexed_sample_dir).rglob("*_metadata.json")):
-                with open(filename_path) as multiplexed_json_file:
-                    multiplexed_json = json.load(multiplexed_json_file)
-
-                library_id = multiplexed_json.pop("library_id")
-                multiplexed_json["scpca_library_id"] = library_id
-                multiplexed_json["scpca_sample_id"] = multiplexed_json.pop("sample_id")
-
-                multiplexed_library_path_mapping[library_id] = multiplexed_sample_dir
-                multiplexed_libraries_metadata.append(multiplexed_json)
-                multiplexed_sample_demux_cell_counter.update(
-                    multiplexed_json["sample_cell_estimates"]
-                )
-
-                # Gather seq_units and technologies data.
-                for demux_sample_id in multiplexed_json["demux_samples"]:
-                    if demux_sample_id not in multiplexed_sample_seq_units_mapping:
-                        multiplexed_sample_seq_units_mapping[demux_sample_id] = set()
-                    if demux_sample_id not in multiplexed_sample_technologies_mapping:
-                        multiplexed_sample_technologies_mapping[demux_sample_id] = set()
-
-                    multiplexed_sample_seq_units_mapping[demux_sample_id].add(
-                        multiplexed_json["seq_unit"].strip()
-                    )
-                    multiplexed_sample_technologies_mapping[demux_sample_id].add(
-                        multiplexed_json["technology"].strip()
-                    )
-
-        combined_single_cell_metadata = self.combine_single_cell_metadata(
-            samples_metadata, single_cell_libraries_metadata, sample_id
-        )
-
-        combined_spatial_metadata = self.combine_spatial_metadata(
-            samples_metadata, spatial_libraries_metadata, sample_id
-        )
-
-        (
-            combined_multiplexed_metadata,
-            multiplexed_sample_mapping,
-        ) = self.combine_multiplexed_metadata(
-            samples_metadata, multiplexed_libraries_metadata, sample_id
-        )
-
-        multiplexed_file_mapping = {
-            ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT: {},
-        }
-        multiplexed_workflow_versions = set()
-        single_cell_file_mapping = {
-            ComputedFile.OutputFileFormats.ANN_DATA: {},
-            ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT: {},
-        }
-        single_cell_workflow_versions = set()
-        spatial_file_mapping = {
-            ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT: {},
-        }
-        spatial_workflow_versions = set()
-
-        samples = self.get_samples(
-            samples_metadata,
-            multiplexed_sample_demux_cell_counter,
-            multiplexed_sample_mapping,
-            multiplexed_sample_seq_units_mapping,
-            multiplexed_sample_technologies_mapping,
-            sample_id=sample_id,
-        )
-
-        def update_ann_data(future):
-            computed_file, metadata_files = future.result()
-            single_cell_file_mapping[ComputedFile.OutputFileFormats.ANN_DATA].update(metadata_files)
-            self.process_computed_file(
-                computed_file, kwargs["clean_up_output_data"], kwargs["update_s3"]
-            )
-
-        def update_multiplexed_data(future):
-            computed_file, metadata_files = future.result()
-            multiplexed_file_mapping[ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT].update(
-                metadata_files
-            )
-            self.process_computed_file(
-                computed_file, kwargs["clean_up_output_data"], kwargs["update_s3"]
-            )
-
-        def update_single_cell_data(future):
-            computed_file, metadata_files = future.result()
-            single_cell_file_mapping[ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT].update(
-                metadata_files
-            )
-            self.process_computed_file(
-                computed_file, kwargs["clean_up_output_data"], kwargs["update_s3"]
-            )
-
-        def update_spatial_data(future):
-            computed_file, metadata_files = future.result()
-            spatial_file_mapping[ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT].update(
-                metadata_files
-            )
-            self.process_computed_file(
-                computed_file, kwargs["clean_up_output_data"], kwargs["update_s3"]
-            )
-
-        max_workers = kwargs["max_workers"]
-        samples_count = len(samples)
-        logger.info(
-            f"Processing {samples_count} sample{pluralize(samples_count)} using "
-            f"{max_workers} worker{pluralize(max_workers)}"
-        )
-        with ThreadPoolExecutor(max_workers=max_workers) as tasks:
-            for sample in Sample.objects.bulk_create(samples):
-                # Skip computed files creation if sample directory does not exist.
-                if sample.scpca_id not in non_downloadable_sample_ids:
-                    libraries = [
-                        library
-                        for library in combined_single_cell_metadata
-                        if library["scpca_sample_id"] == sample.scpca_id
-                    ]
-                    workflow_versions = [library["workflow_version"] for library in libraries]
-                    single_cell_workflow_versions.update(workflow_versions)
-
-                    file_formats = [ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT]
-                    if sample.includes_anndata:
-                        file_formats.append(ComputedFile.OutputFileFormats.ANN_DATA)
-
-                    for file_format in file_formats:
-                        tasks.submit(
-                            ComputedFile.get_sample_single_cell_file,
-                            sample,
-                            libraries,
-                            workflow_versions,
-                            file_format,
-                        ).add_done_callback(
-                            update_single_cell_data
-                            if file_format == ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT
-                            else update_ann_data
-                        )
-
-                        if sample.has_spatial_data:
-                            libraries = [
-                                library
-                                for library in combined_spatial_metadata
-                                if library["scpca_sample_id"] == sample.scpca_id
-                            ]
-                            workflow_versions = [
-                                library["workflow_version"] for library in libraries
-                            ]
-                            spatial_workflow_versions.update(workflow_versions)
-                            tasks.submit(
-                                ComputedFile.get_sample_spatial_file,
-                                sample,
-                                libraries,
-                                workflow_versions,
-                                ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT,
-                            ).add_done_callback(update_spatial_data)
-
-                if sample.has_multiplexed_data:
-                    libraries = [
-                        library
-                        for library in combined_multiplexed_metadata
-                        if library.get("scpca_sample_id") == sample.scpca_id
-                    ]
-                    workflow_versions = [library["workflow_version"] for library in libraries]
-                    multiplexed_workflow_versions.update(workflow_versions)
-                    tasks.submit(
-                        ComputedFile.get_sample_multiplexed_file,
-                        sample,
-                        libraries,
-                        multiplexed_library_path_mapping,
-                        workflow_versions,
-                        ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT,
-                    ).add_done_callback(update_multiplexed_data)
-
-        if multiplexed_file_mapping[ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT]:
-            # We want a single ZIP archive for a multiplexed samples project.
-            multiplexed_file_mapping[ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT].update(
-                single_cell_file_mapping[ComputedFile.OutputFileFormats.SINGLE_CELL_EXPERIMENT]
-            )
-        self.create_computed_files(
-            single_cell_file_mapping,
-            single_cell_workflow_versions,
-            spatial_file_mapping,
-            spatial_workflow_versions,
-            multiplexed_file_mapping,
-            multiplexed_workflow_versions,
-            clean_up_output_data=kwargs["clean_up_output_data"],
-            update_s3=kwargs["update_s3"],
-        )
-
-        # Set modality flags based on a real data availability.
-        self.has_bulk_rna_seq = self.samples.filter(has_bulk_rna_seq=True).exists()
-        self.has_cite_seq_data = self.samples.filter(has_cite_seq_data=True).exists()
-        self.has_multiplexed_data = self.samples.filter(has_multiplexed_data=True).exists()
-        self.has_single_cell_data = self.samples.filter(has_single_cell_data=True).exists()
-        self.has_spatial_data = self.samples.filter(has_spatial_data=True).exists()
-        self.includes_anndata = self.samples.filter(includes_anndata=True).exists()
-        self.save(
-            update_fields=(
-                "has_bulk_rna_seq",
-                "has_cite_seq_data",
-                "has_multiplexed_data",
-                "has_single_cell_data",
-                "has_spatial_data",
-                "includes_anndata",
-            )
-        )
-
-        self.update_counts()
-
+    # TODO - 'purge from db' here, move 'purge from s3' elsewhere?
     def purge(self, delete_from_s3=False):
         """Purges project and its related data."""
         for sample in self.samples.all():
